+++ projectroles/templatetags/projectroles_tags.py
@@ -130,42 +130,6 @@ def get_project_type_str(project, capitalize=True):
 
 
 @register.simple_tag
-def highlight_search_term(item, term):
-    """Return string with search term highlighted"""
-
-    def get_highlights(item):
-        pos = item.lower().find(term)
-        tl = len(term)
-
-        if pos == -1:
-            return item     # Nothing to highlight
-
-        ret = item[:pos]
-        ret += '<span class="omics-search-highlight">' + \
-               item[pos:pos + tl] + '</span>'
-
-        if len(item[pos + tl:]) > 0:
-            ret += get_highlights(item[pos + tl:])
-
-        return ret
-
-    return get_highlights(item)
-
-
-@register.simple_tag
-def get_project_title_html(project):
-    """Return HTML version of the full project title including parents"""
-    ret = ''
-
-    if project.get_parents():
-        ret += '<span class="text-muted">{}</span>'.format(
-            ' / '.join(project.get_full_title().split(' / ')[:-1]) + ' / ')
-
-    ret += project.title
-    return ret
-
-
-@register.simple_tag
 def get_star(project, user):
     """Return HTML for project star tag state if it is set"""
     if (user.has_perm('projectroles.view_project', project) and