+++ timeline/views.py
@@ -32,6 +32,15 @@ class ProjectTimelineView(
         except Project.DoesNotExist:
             return None
 
+    def get_context_data(self, *args, **kwargs):
+        context = super(
+            ProjectTimelineView, self).get_context_data(*args, **kwargs)
+
+        context['timeline_title'] = 'Project Timeline'
+        context['timeline_mode'] = 'project'
+
+        return context
+
     def get_queryset(self):
         set_kwargs = {
             'project': self.kwargs['project']}
@@ -43,6 +52,33 @@ class ProjectTimelineView(
 
         return ProjectEvent.objects.filter(
             **set_kwargs).order_by('-pk')
+
+
+class ObjectTimelineView(ProjectTimelineView):
+    """View for displaying files and folders for a project"""
+
+    def get_context_data(self, *args, **kwargs):
+        context = super(
+            ObjectTimelineView, self).get_context_data(*args, **kwargs)
+
+        context['timeline_title'] = '{} Timeline'.format(
+            self.kwargs['object_model'])
+        context['timeline_mode'] = 'object'
+
+        return context
+
+    def get_queryset(self):
+        queryset = ProjectEvent.objects.get_object_events(
+            project_pk=self.kwargs['project'],
+            object_model=self.kwargs['object_model'],
+            object_pk=self.kwargs['object_pk'])
+
+        if not self.request.user.has_perm(
+                'timeline.view_classified_event',
+                self.get_permission_object()):
+            queryset = queryset.filter(classified=False)
+
+        return queryset
 
 
 # Taskflow API Views -----------------------------------------------------